% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regiaoDoc.R
\docType{data}
\name{regiao}
\alias{regiao}
\title{Dados territoriais do SUS}
\format{An object of class \code{data.frame} with 5631 rows and 30 columns.}
\usage{
regiao
}
\description{
Banco de dados com municípios e regiões administrativas e de saúde brasileiras usadas pelo DATASUS
}
\keyword{datasets}

