% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsaude.R
\name{regional}
\alias{regional}
\title{Nomeia as regiões de saúde}
\usage{
regional(x)
}
\arguments{
\item{x}{objeto da classe `data.frame` com pelo menos uma variável, de nome `CO_MUNICIP`, com o código IBGE do município}
}
\description{
Inclui o nome da Região de Saúde do município numa base de dados com municípios brasileiros identificados pelo código do IBGE em variável de nome `CO_MUNICIP`.
}

