% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territorio-package.R
\docType{package}
\name{territorio-package}
\alias{territorio-package}
\title{territorio: Brazilian administrative and health divisions}
\description{
The territorio package provides a data set with the Brazilian municipalities, their bounderies in geographic coordinates and the administrative and health division to which municipality belongs, and a function to subset the municipalities by State.
}
\section{Package objects}{

 \itemize{
   \item Dataset \code{regiao}
   \item Function \code{territorio}
 }
}
\author{
\packageAuthor{territorio} <fulvio.nedel@ufsc.br>
}

