% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territorio.R
\name{territorio}
\alias{territorio}
\title{Classifica os municípios nas unidades administrativas e de saúde brasileiras}
\usage{
territorio(local = "SC", ano = 2012, tipo = "curto")
}
\arguments{
\item{local}{"SC" (padrão) seleciona apenas os municípios de Santa Catarina, "RS" seleciona apenas os municípios do Rio Grande do Sul. Qualquer outra opção (deve ir entre aspas, simples ou duplas) retorna um banco com todos os municípios brasileiros.}

\item{ano}{Exclui municípios com data de instalação posterior ao ano de escolha. O padrão é 2012.}

\item{tipo}{"curto" (padrão) retorna um banco com oito variáveis, qualquer outra opção (deve ir entre aspas, simples ou duplas) retorna um banco com 30 variáveis}
}
\description{
Classifica os municípios brasileiros em divisões administrativas e de saúde segundo as listas do DATASUS
}
\note{
O arquivo na "Transferência de arquivos" ou no servidor `ftp` do DATASUS fala em "base territorial 2005", mas inclui um município criado em 2009 e cinco em 2013. Esses municípios podem ser excluídos com o argumento `ano` (v. `Arguments`).

No momento a função tá pensada pra trabalhar com os municípios de Santa Catarina em 2012.
}

